requirements.txt
Flask==3.1.0
Django==4.0.3
bottle==0.13.4
python-decouple==3.6
django-cors-headers==3.11.0
djangorestframework==3.13.1